Remove process information from iOS backend
iOS apps are sandboxed and unable to fetch this information at all
