+++ src/apple/ios/mod.rs
@@ -6,3 +6,4 @@
 
 pub mod component;
 pub mod ffi {}
+pub mod process;
