+++ src/apple/macos/ffi.rs
@@ -10,11 +10,31 @@ use core_foundation_sys::dictionary::CFMutableDictionaryRef;
 use core_foundation_sys::string::{CFStringEncoding, CFStringRef};
 use core_foundation_sys::url::CFURLRef;
 
-use libc::{c_char, c_void, size_t};
+use libc::{c_char, c_int, c_void, size_t};
 
 pub(crate) use crate::sys::ffi::*;
 
 extern "C" {
+    pub fn mach_absolute_time() -> u64;
+
+    pub fn proc_pidinfo(
+        pid: c_int,
+        flavor: c_int,
+        arg: u64,
+        buffer: *mut c_void,
+        buffersize: c_int,
+    ) -> c_int;
+
+    pub fn proc_listallpids(buffer: *mut c_void, buffersize: c_int) -> c_int;
+    //pub fn proc_listpids(kind: u32, x: u32, buffer: *mut c_void, buffersize: c_int) -> c_int;
+    //pub fn proc_name(pid: c_int, buffer: *mut c_void, buffersize: u32) -> c_int;
+    //pub fn proc_regionfilename(pid: c_int, address: u64, buffer: *mut c_void,
+    //                           buffersize: u32) -> c_int;
+
+    pub fn proc_pidpath(pid: c_int, buffer: *mut c_void, buffersize: u32) -> c_int;
+
+    pub fn proc_pid_rusage(pid: c_int, flavor: c_int, buffer: *mut c_void) -> c_int;
+
     // IOKit is only available on MacOS: https://developer.apple.com/documentation/iokit
 
     pub fn IOMasterPort(a: i32, b: *mut mach_port_t) -> i32;
@@ -143,3 +163,8 @@ pub const SMC_CMD_READ_KEYINFO: u8 = 9;
 pub const SMC_CMD_READ_BYTES: u8 = 5;
 
 pub const KIO_RETURN_SUCCESS: i32 = 0;
+
+//pub const PROC_ALL_PIDS: c_uint = 1;
+pub const PROC_PIDTBSDINFO: c_int = 3;
+
+pub const PROC_PIDPATHINFO_MAXSIZE: u32 = 4096;
