+++ src/apple/system.rs
@@ -13,16 +13,20 @@ use crate::sys::processor::*;
 #[cfg(target_os = "macos")]
 use core_foundation_sys::base::{kCFAllocatorDefault, CFRelease};
 
-use crate::utils::into_iter;
+use crate::{LoadAvg, Pid, ProcessorExt, RefreshKind, SystemExt, User};
 
-use crate::{LoadAvg, Pid, ProcessExt, ProcessorExt, RefreshKind, SystemExt, User};
+#[cfg(target_os = "macos")]
+use crate::ProcessExt;
 
 use std::cell::UnsafeCell;
 use std::collections::HashMap;
 use std::mem;
 use std::sync::Arc;
 
-use libc::{self, c_char, c_int, c_void, size_t, sysconf, _SC_PAGESIZE};
+#[cfg(target_os = "macos")]
+use libc::size_t;
+
+use libc::{self, c_char, c_int, c_void, sysconf, _SC_PAGESIZE};
 
 /// Structs containing system's information.
 pub struct System {
@@ -73,12 +77,15 @@ pub(crate) struct Wrap<'a>(pub UnsafeCell<&'a mut HashMap<Pid, Process>>);
 unsafe impl<'a> Send for Wrap<'a> {}
 unsafe impl<'a> Sync for Wrap<'a> {}
 
+#[cfg(target_os = "macos")]
 impl System {
     fn clear_procs(&mut self) {
+        use crate::sys::macos::process;
+
         let mut to_delete = Vec::new();
 
         for (pid, mut proc_) in &mut self.process_list {
-            if !has_been_updated(&mut proc_) {
+            if !process::has_been_updated(&mut proc_) {
                 to_delete.push(*pid);
             }
         }
@@ -269,7 +276,13 @@ impl SystemExt for System {
             .set_cpu_usage(pourcent / self.processors.len() as f32);
     }
 
+    #[cfg(target_os = "ios")]
+    fn refresh_processes(&mut self) {}
+
+    #[cfg(target_os = "macos")]
     fn refresh_processes(&mut self) {
+        use crate::utils::into_iter;
+
         let count = unsafe { ffi::proc_listallpids(::std::ptr::null_mut(), 0) };
         if count < 1 {
             return;
@@ -296,6 +309,12 @@ impl SystemExt for System {
         }
     }
 
+    #[cfg(target_os = "ios")]
+    fn refresh_process(&mut self, _: Pid) -> bool {
+        false
+    }
+
+    #[cfg(target_os = "macos")]
     fn refresh_process(&mut self, pid: Pid) -> bool {
         let arg_max = get_arg_max();
         match {
@@ -524,6 +543,7 @@ fn get_io_service_connection() -> Option<ffi::io_connect_t> {
     }
 }
 
+#[cfg(target_os = "macos")]
 fn get_arg_max() -> usize {
     let mut mib: [c_int; 3] = [libc::CTL_KERN, libc::KERN_ARGMAX, 0];
     let mut arg_max = 0i32;
